Add files via upload
This project in RobotStudio involves 10 different boxes and 10 different containers, where each container corresponds to a specific box. The system receives commands via a UDP interface, specifying the box number and the orientation in which it should be picked up. The item is then placed in its corresponding container, but with the default orientation.